Tying up loose ends...
