+++ GlobalHotkey.cs
@@ -9,7 +9,6 @@ namespace MouseNet.Logophi
     {
         private const int HotkeyMsg = 0x0312;
         private readonly Form _form;
-        private int _currentId;
         private readonly IList<Keys> _hotkeys = new List<Keys>();
 
         public GlobalHotkey()
@@ -18,11 +17,11 @@ namespace MouseNet.Logophi
             }
 
         public IEnumerable<Keys> Hotkeys => _hotkeys;
-        public int HotkeyCount => _currentId;
+        public int HotkeyCount { get; private set; }
 
         public void Dispose()
             {
-            for (var i = 0; i < _currentId; i++)
+            for (var i = 0; i < HotkeyCount; i++)
                 UnregisterHotkey(i);
             _form?.Dispose();
             }
@@ -34,13 +33,13 @@ namespace MouseNet.Logophi
                 (int) hotkey.GetModifiers().ToModifierKeys();
             var keycode = (int) hotkey.GetKeyCode();
             if (!NativeMethods.RegisterHotKey(_form.Handle,
-                                              _currentId++,
+                                              HotkeyCount++,
                                               modifiers,
                                               keycode))
                 throw new InvalidOperationException(
                     "Failed to register hotkey.");
             _hotkeys.Add(hotkey);
-            return _currentId;
+            return HotkeyCount;
             }
 
         public void UnregisterHotkey
@@ -85,65 +84,5 @@ namespace MouseNet.Logophi
                     new HotkeyEventArgs(modifiers.ToKeys(), hotkey));
                 }
         }
-    }
-
-    public class HotkeyEventArgs : EventArgs
-    {
-        internal HotkeyEventArgs
-            (Keys modifiers,
-             Keys key)
-            {
-            Modifiers = modifiers;
-            KeyCode = key;
-            }
-
-        public Keys KeyCode { get; }
-        public Keys Modifiers { get; }
-        public Keys Hotkey => KeyCode | Modifiers;
-    }
-
-    public static class Extensions
-    {
-        public static Keys GetKeyCode
-            (this Keys keys)
-            {
-            return keys & Keys.KeyCode;
-            }
-
-        public static Keys GetModifiers
-            (this Keys keys)
-            {
-            return keys & Keys.Modifiers;
-            }
-
-        public static Keys ToKeys
-            (this ModifierKeys modifierKeys)
-            {
-            var key = Keys.None;
-            if (modifierKeys.HasFlag(ModifierKeys.Alt))
-                key |= Keys.Alt;
-            if (modifierKeys.HasFlag(ModifierKeys.Control))
-                key |= Keys.Control;
-            if (modifierKeys.HasFlag(ModifierKeys.Shift))
-                key |= Keys.Shift;
-            if (modifierKeys.HasFlag(ModifierKeys.Windows))
-                key |= Keys.LWin;
-            return key;
-            }
-
-        public static ModifierKeys ToModifierKeys
-            (this Keys keys)
-            {
-            var modifiers = ModifierKeys.None;
-            if (keys.HasFlag(Keys.Alt))
-                modifiers |= ModifierKeys.Alt;
-            if (keys.HasFlag(Keys.Control))
-                modifiers |= ModifierKeys.Control;
-            if (keys.HasFlag(Keys.Shift))
-                modifiers |= ModifierKeys.Shift;
-            if (keys.HasFlag(Keys.LWin))
-                modifiers |= ModifierKeys.Windows;
-            return modifiers;
-            }
     }
 }
