+++ Logophi.csproj
@@ -110,6 +110,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AppContext.cs" />
+    <Compile Include="Extensions.cs" />
     <Compile Include="Forms\AboutForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -123,6 +124,7 @@
       <DependentUpon>BookmarksForm.cs</DependentUpon>
     </Compile>
     <Compile Include="GlobalHotkey.cs" />
+    <Compile Include="HotkeyEventArgs.cs" />
     <Compile Include="IBookmarkManager.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="PresentationAgent.cs" />
